requirements.txt
twitchio==3.2.1
python-dotenv==1.2.2
